requirements.txt
blinker==1.9.0
boto3==1.42.66
botocore==1.42.66
click==8.1.8
Flask==3.1.3
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
MarkupSafe==3.0.3
pillow==11.3.0
pyobjc-core==12.0
pyobjc-framework-Cocoa==12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.16.0
six==1.17.0
urllib3==1.26.20
Werkzeug==3.1.6
zipp==3.23.0
